requirements.txt
pytest
allure-pytest
requests
selenium
pytest-xdist